Added function to add a background cmpnt to a window
NOTE PRESSING STUFF WONT WORK FOR A BIT

+++ Project-Perfect-Citizen/Code/Game/createWindow.cpp
@@ -26,7 +26,7 @@
 
 using namespace ppc;
 
-WindowInterface* ppc::spawnConsole(InputHandler & ih, NodeState & ns, int x, int y, int w, int h) {
+WindowInterface* ppc::spawnConsole(InputHandler & ih, NodeState & ns, float x, float y, int w, int h) {
 
 	
 	/////// COMPONENTS /////////
